+++ indicators-dao/src/main/java/org/esco/indicators/dao/statistic/PortalConnectionStatisticDaoImpl.java
@@ -67,7 +67,7 @@ public class PortalConnectionStatisticDaoImpl implements PortalConnectionStatist
 	parameters.put("userProfile", userProfile);
 	
 	// Retrieval of the result
-	Long result = (Long) QueryManager.getSingleResult(entityManager, namedQuery, parameters);
+	Double result = (Double) QueryManager.getSingleResult(entityManager, namedQuery, parameters);
 	Float averageDurationTime = (result != null ? result.floatValue() : null);
 	
 	return averageDurationTime;
@@ -211,7 +211,7 @@ public class PortalConnectionStatisticDaoImpl implements PortalConnectionStatist
 	parameters.put("userProfile", userProfile);
 	
 	// Retrieval of the result
-	Long result = (Long) QueryManager.getSingleResult(entityManager, namedQuery, parameters);
+	Double result = (Double) QueryManager.getSingleResult(entityManager, namedQuery, parameters);
 	Float averageDurationTime = (result != null ? result.floatValue() : null);
 	
 	return averageDurationTime;
